+++ variables.tf
@@ -75,6 +75,7 @@ variable "drupal_projects_list" {
     bucket_obj_vwr                       = optional(list(string), [])
     bucket_soft_delete_retention_seconds = optional(number, 0)
     network_policy                       = optional(string, "")
+    network_policy_acme_port             = optional(number, 8089)
   }))
 
   validation {